+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -41864,6 +41864,74 @@ $ORG_SelectMap = @{
 }
 
 _awsArgumentCompleterRegistration $ORG_SelectCompleters $ORG_SelectMap
+# Argument completions for service Amazon OpenSearch Ingestion
+
+
+$OSIS_SelectCompleters = {
+    param($commandName, $parameterName, $wordToComplete, $commandAst, $fakeBoundParameter)
+
+    $cmdletType = Invoke-Expression "[Amazon.PowerShell.Cmdlets.OSIS.$($commandName.Replace('-', ''))Cmdlet]"
+    if (-not $cmdletType) {
+        return
+    }
+    $awsCmdletAttribute = $cmdletType.GetCustomAttributes([Amazon.PowerShell.Common.AWSCmdletAttribute], $false)
+    if (-not $awsCmdletAttribute) {
+        return
+    }
+    $type = $awsCmdletAttribute.SelectReturnType
+    if (-not $type) {
+        return
+    }
+
+    $splitSelect = $wordToComplete -Split '\.'
+    $splitSelect | Select-Object -First ($splitSelect.Length - 1) | ForEach-Object {
+        $propertyName = $_
+        $properties = $type.GetProperties(('Instance', 'Public', 'DeclaredOnly')) | Where-Object { $_.Name -ieq $propertyName }
+        if ($properties.Length -ne 1) {
+            break
+        }
+        $type = $properties.PropertyType
+        $prefix += "$($properties.Name)."
+
+        $asEnumerableType = $type.GetInterface('System.Collections.Generic.IEnumerable`1')
+        if ($asEnumerableType -and $type -ne [System.String]) {
+            $type =  $asEnumerableType.GetGenericArguments()[0]
+        }
+    }
+
+    $v = @( '*' )
+    $properties = $type.GetProperties(('Instance', 'Public', 'DeclaredOnly')).Name | Sort-Object
+    if ($properties) {
+        $v += ($properties | ForEach-Object { $prefix + $_ })
+    }
+    $parameters = $cmdletType.GetProperties(('Instance', 'Public')) | Where-Object { $_.GetCustomAttributes([System.Management.Automation.ParameterAttribute], $true) } | Select-Object -ExpandProperty Name | Sort-Object
+    if ($parameters) {
+        $v += ($parameters | ForEach-Object { "^$_" })
+    }
+
+    $v |
+        Where-Object { $_ -match "^$([System.Text.RegularExpressions.Regex]::Escape($wordToComplete)).*" } |
+        ForEach-Object { New-Object System.Management.Automation.CompletionResult $_, $_, 'ParameterValue', $_ }
+}
+
+$OSIS_SelectMap = @{
+    "Select"=@("New-OSISPipeline",
+               "Remove-OSISPipeline",
+               "Get-OSISPipeline",
+               "Get-OSISPipelineBlueprint",
+               "Get-OSISPipelineChangeProgress",
+               "Get-OSISPipelineBlueprintList",
+               "Get-OSISPipelineList",
+               "Get-OSISResourceTag",
+               "Start-OSISPipeline",
+               "Stop-OSISPipeline",
+               "Add-OSISResourceTag",
+               "Remove-OSISResourceTag",
+               "Update-OSISPipeline",
+               "Use-OSISPipeline")
+}
+
+_awsArgumentCompleterRegistration $OSIS_SelectCompleters $OSIS_SelectMap
 # Argument completions for service AWS Outposts