+++ modules/AWSPowerShell/Cmdlets/OSIS/AWS.Tools.OSIS.Aliases.psm1
@@ -1,0 +1,3 @@
+﻿# OSIS
+
+Export-ModuleMember -Alias *